refactor(api): restructure Pochi API types and client implementations
- Move Pochi API type definitions from common package to vendor-pochi package
- Remove unused PersistRequest/PersistResponse types (using webhook instead)
- Simplify attachment upload implementation in vscode-webui
- Remove unused ApiClient from vscode extension
- Clean up auth-client implementation in vscode package
