+++ packages/vscode-webui/src/lib/hooks/use-attachment-upload.ts
@@ -1,9 +1,7 @@
 import { MaxAttachments } from "@/lib/constants";
 import { createFileName, validateFile } from "@/lib/utils/attachment";
-import type { PochiApi } from "@getpochi/common/pochi-api";
 import { getServerBaseUrl } from "@getpochi/common/vscode-webui-bridge";
 import type { FileUIPart } from "ai";
-import { hc } from "hono/client";
 import { useRef, useState } from "react";
 
 interface UseAttachmentUploadOptions {
@@ -206,19 +204,14 @@ export function useAttachmentUpload(options?: UseAttachmentUploadOptions) {
 // }
 
 async function fileToRemoteUri(file: File, signal?: AbortSignal) {
-  const apiClient = hc<PochiApi>(getServerBaseUrl());
-  const response = await apiClient.api.upload.$post(
-    {
-      form: {
-        file,
-      },
-    },
-    {
-      init: {
-        signal,
-      },
-    },
-  );
+  const formData = new FormData();
+  formData.append("file", file);
+
+  const response = await fetch(`${getServerBaseUrl()}/api/upload`, {
+    method: "POST",
+    body: formData,
+    signal,
+  });
 
   if (!response.ok) {
     throw new Error(`Upload failed: ${response.statusText}`);
